Address quantity description review feedback
